python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bedrock-agentcore[strands-agents]>=1.1.2",
    "botocore[crt]>=1.42.17",
    "strands-agents[otel]>=1.20.0",
    "strands-agents-tools>=0.2.18",
    "bedrock-agentcore-starter-toolkit>=0.2.5",
]
